refactor: lib -> utils
